update cpu.c, mm.c, os.c, timer.c

+++ ossim_source_code/src/cpu.c
@@ -4,7 +4,8 @@
 #include "mm.h"
 
 int calc(struct pcb_t * proc) {
-	return ((unsigned long)proc & 0UL);
+	// return ((unsigned long)proc & 0UL);
+	return ((uintptr_t)proc & 0UL);
 }
 
 int alloc(struct pcb_t * proc, uint32_t size, uint32_t reg_index) {
